make simple bar chart tool tip it's own component
