+++ app/components/ChartSimpleBar.js
@@ -2,6 +2,8 @@ import { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, Legend, Responsive
 var React = require('react');
 var PropTypes = React.PropTypes;
 
+var CustomToolTipSimpleBarCharts = require('./CustomToolTipSimpleBarCharts')
+
 import {
   BOX_BORDER,
   SPACING,
@@ -65,73 +67,6 @@ const ChartSimpleBar = React.createClass({
 
     const self = this;
 
-    const tooltipstyle = {
-      width: '100%',
-      margin: 0,
-      lineHeight: 24,
-      border: '1px solid #f5f5f5',
-      backgroundColor: 'rgba(255, 255, 255, 0.95)',
-      padding: 10,
-    };
-
-    const toolTipLabel = {
-      margin: '0',
-      color: '#666',
-      fontWeight: '700',
-    };
-
-    const CustomTooltip  = React.createClass({
-      propTypes: {
-        type: PropTypes.string,
-        payload: PropTypes.array,
-        label: PropTypes.string,
-      },
-
-      render() {
-        const { active } = this.props;
-        let html_hov = '';
-        if (active) {
-          const { payload, label } = this.props;
-
-          const thedata = payload.map( bar_segment => {
-
-           const colors = self.get_keyColors(bar_segment.name)
-
-            const toolTipName = {
-              margin: '0',
-              color: colors,
-            }
-
-            const toolTipValue = {
-              fontWeight: '800',
-            }
-
-            const value = bar_segment.value ?  bar_segment.value.toString().substring(0,5) : 'N/A';
-            const name = bar_segment.name + ": "
-
-            return ( <p key={bar_segment.name} style={toolTipName}>{name}<span style={toolTipValue}>{value}</span></p>)
-          })
-
-          if (label === '1' || label === '2' ){
-            return (<div key={label+'blanktip'}/>)
-          }
-
-          const labelstr = label.toString();
-
-          return (
-
-            <div key={labelstr+'tooltip'} style={tooltipstyle}>
-              <p style={toolTipLabel}>Cathcment: {`${labelstr}`}</p>
-              {thedata}
-          </div>
-
-          );
-        }
-
-        return null;
-      }
-    });
-
     const data = this.props.chart_data
     const bars = this.get_bars(data[0])
     const datas = data[0]
@@ -178,7 +113,7 @@ const ChartSimpleBar = React.createClass({
                 margin={{top: 5, right: 5, left: 5, bottom: 5}}>
                 <XAxis dataKey="name" hide={false} tick={false} tickLine={false} axisLine={false} />
                 <YAxis hide={false} tick={false} tickLine={false} axisLine={false} />
-                <Tooltip content={<CustomTooltip />}/>
+                <Tooltip content={<CustomToolTipSimpleBarCharts  get_keyColors={this.get_keyColors} />}/>
                 <Legend />
                 {bars}
               </BarChart>
@@ -190,9 +125,5 @@ const ChartSimpleBar = React.createClass({
     );
   }
 })
-//{bars}
 
-// <Bar key="Habitat" dataKey="Habitat"  fill={this.get_keyColors("Habitat")}  />
-// <Bar key="Hydrology" dataKey="Hydrology"  fill={this.get_keyColors("Hydrology")}  />
-// <Bar key="Water Quality" dataKey="Water Quality"  fill={this.get_keyColors("Water Quality")}  />
 module.exports = ChartSimpleBar;
